rename the submodule `model` to `models`
