+++ QGrain.pro
@@ -1,57 +1,56 @@
 
-SOURCES=C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\__main__.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\artificial\_generator.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\artificial\_preset.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\BaseChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\BoxplotChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\config_matplotlib.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\CumulativeCurveChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\diagrams.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\DistanceCurveChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\DistributionChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\EMMAResultChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\Frequency3DChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\FrequencyCurveChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\HierarchicalChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\PCAResultChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\chart\UDMResultChart.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\emma\_distance.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\emma\_kernel.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\emma\_resolver.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\emma\_result.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\emma\_setting.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\io\_load.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\io\_save.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\io\_use_excel.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\model\_dataset.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ssu\_distance.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ssu\_distribution.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ssu\_resolver.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ssu\_result.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ssu\_task.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\statistics\_base.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\statistics\_GRADISTAT.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\udm\_resolver.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\udm\_result.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\udm\_setting.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\AboutDialog.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\ClusteringAnalyzer.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\DatasetGenerator.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\EMMAAnalyzer.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\EMMASettingDialog.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\GrainSizeDatasetViewer.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\LoadDatasetDialog.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\LogDialog.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\MainWindow.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\ParameterEditor.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\ParameterTable.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\PCAAnalyzer.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\SSUAnalyzer.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\SSUMulticoreAnalyzer.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\SSUReferenceViewer.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\SSUResultViewer.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\SSUSettingDialog.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\UDMAnalyzer.py\
-    C:\Users\yurio\OneDrive\Codes\QGrain\QGrain\ui\UDMSettingDialog.py
-TRANSLATIONS=en.ts zh_CN.ts
+SOURCES=.\QGrain\statistics.py\
+    .\QGrain\__main__.py\
+    .\QGrain\artificial\_generator.py\
+    .\QGrain\artificial\_preset.py\
+    .\QGrain\chart\BaseChart.py\
+    .\QGrain\chart\BoxplotChart.py\
+    .\QGrain\chart\config_matplotlib.py\
+    .\QGrain\chart\CumulativeCurveChart.py\
+    .\QGrain\chart\diagrams.py\
+    .\QGrain\chart\DistanceCurveChart.py\
+    .\QGrain\chart\DistributionChart.py\
+    .\QGrain\chart\EMMAResultChart.py\
+    .\QGrain\chart\Frequency3DChart.py\
+    .\QGrain\chart\FrequencyCurveChart.py\
+    .\QGrain\chart\HierarchicalChart.py\
+    .\QGrain\chart\PCAResultChart.py\
+    .\QGrain\chart\UDMResultChart.py\
+    .\QGrain\emma\_distance.py\
+    .\QGrain\emma\_kernel.py\
+    .\QGrain\emma\_resolver.py\
+    .\QGrain\emma\_result.py\
+    .\QGrain\emma\_setting.py\
+    .\QGrain\io\_load.py\
+    .\QGrain\io\_save.py\
+    .\QGrain\io\_use_excel.py\
+    .\QGrain\models\dataset.py\
+    .\QGrain\ssu\_distance.py\
+    .\QGrain\ssu\_distribution.py\
+    .\QGrain\ssu\_resolver.py\
+    .\QGrain\ssu\_result.py\
+    .\QGrain\ssu\_task.py\
+    .\QGrain\udm\_resolver.py\
+    .\QGrain\udm\_result.py\
+    .\QGrain\udm\_setting.py\
+    .\QGrain\ui\AboutDialog.py\
+    .\QGrain\ui\ClusteringAnalyzer.py\
+    .\QGrain\ui\DatasetGenerator.py\
+    .\QGrain\ui\EMMAAnalyzer.py\
+    .\QGrain\ui\EMMASettingDialog.py\
+    .\QGrain\ui\GrainSizeDatasetViewer.py\
+    .\QGrain\ui\LoadDatasetDialog.py\
+    .\QGrain\ui\LogDialog.py\
+    .\QGrain\ui\MainWindow.py\
+    .\QGrain\ui\ParameterEditor.py\
+    .\QGrain\ui\ParameterTable.py\
+    .\QGrain\ui\PCAAnalyzer.py\
+    .\QGrain\ui\SSUAnalyzer.py\
+    .\QGrain\ui\SSUMulticoreAnalyzer.py\
+    .\QGrain\ui\SSUReferenceViewer.py\
+    .\QGrain\ui\SSUResultViewer.py\
+    .\QGrain\ui\SSUSettingDialog.py\
+    .\QGrain\ui\UDMAnalyzer.py\
+    .\QGrain\ui\UDMSettingDialog.py
+TRANSLATIONS=zh_CN.xml
 CODECFORTR=UTF-8
